+++ CHANGELOG.rst
@@ -4,6 +4,18 @@ Changelog
 Features
 ^^^^^^^^
 
+* HTTP(S) (``http``)
+
+  * Headers (``headers``)
+
+    * add parsers for security related headers (\
+      `Expect-CT <https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Expect-CT>`_, \
+      `Expect-Staple <https://scotthelme.co.uk/designing-a-new-security-header-expect-staple>`_, \
+      `Strict-Transport-Security <https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Strict-Transport-Security>`_, \
+      `X-Content-Type-Options <https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Content-Type-Options>`_, \
+      `X-Frame-Options <https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Frame-Options>`_)
+
+
 * TLS (``tls``)
 
   * Versions (``versions``)